+++ Changelog.md
@@ -1,6 +1,23 @@
 # Versions:
 
+## 1.0.2 (25.11.2024)
+
+---
+
+### Backend:
+
+- Added more logs
+- ``username`` is not unique anymore, only ``email`` is still unique.
+- Removed redundant PATCH function ``/modifiy/password``.
+- ``Roles.initRoles`` ignores default roles ``ADMIN`` and ``USER``
+- Added documentation (openapi.json, redoc-static.html, client.http)
+
+### Frontend:
+- No changes
+
 ## 1.0.1 (08.11.2024)
+
+---
 
 ### Backend: